feat: HearthDemControl, the shared DEM surface helper
Task 5 of energy round C1 (design spec 4.1): the shared
DeviceEnergyManagement (0x0098, cluster 152) helper, built fresh in the
HearthMeasurementPush pattern (owner pointer, enabled flag,
onReconciled(), reset()), for Task 6's MatterBatteryStorage (FULL
variant) and MatterDeviceEnergyManagement (both variants) to embed. The
public surface is the design spec's class block verbatim: PowerAdjustEntry,
the six scalar setters, pushAdjustmentEnergyUse (field 6, event carrier),
setPowerAdjustmentCapability (AT+MTDEMCAP), onPowerAdjust/
onCancelPowerAdjust as plain function pointers (a deliberate departure
from this library's usual std::function convention, followed exactly as
written), endAdjustment(), onReconciled(), enabled.

Four semantics honoured and test-pinned: an accepted PowerAdjustRequest
pushes no state change (the firmware owns the transition), but silently
arms the internal ESAState cache so a later endAdjustment() reaches the
wire honestly instead of no-opping against a stale belief; field 6 keeps
no cache and always writes, unlike every other field's null-until-pushed
discipline; the reconcile split re-pushes configuration (ESAType,
ESACanGenerate, AbsMin/MaxPower, the capability list) and clears-without-
resending the volatile fields (ESAState, OptOutState); enabled=false
refuses every wire-touching call host-side with zero traffic, a wider
gate than HearthMeasurementPush's (which only gates its energy adders).

Beyond the brief's file list, disclosed and precedented by 0.9.0's own
four-to-five +MTCMD field growth for Boost: HearthCmdFields.value[]
(MatterEndPoint.h) widened from uint32_t to int64_t, and
hearthDispatchCmd()'s tail parse (Hearth.cpp) switched from strtoul() to
HearthCompat.h's hearthParseWireValue() (already used for +MTATTR's
identical width problem). PowerAdjustRequest's power field is genuinely
int64 mW on the wire, and the design spec's harness names 5000000000
(>2^32) as the canonical bench vector; the previous uint32_t field would
have silently truncated it before it ever reached a sketch's callback.
Verified behaviour-preserving for every existing +MTCMD consumer (all
already narrow their own read of value[i]) before making the change.

Host suite: 3744 passed, 0 failed, across 54 binaries (test_demcontrol
new at 113 checks; the 53 pre-existing RESULT-convention binaries plus
test_transport's own convention are otherwise byte-for-byte unchanged).

+++ src/MatterEndPoint.h
@@ -69,12 +69,28 @@
  * mixed `present` array; a legacy line that never reaches a given position
  * at all leaves it both absent and, since the struct is zero-initialised by
  * hearthDispatchCmd() before parsing, value 0.
+ *
+ * Energy round C1 (Task 5, HearthDemControl's PowerAdjustRequest forward,
+ * AT_MT_SPEC.md S3.17) widened `value[]` from uint32_t to int64_t, and
+ * hearthDispatchCmd()'s tail parse from strtoul() to HearthCompat.h's
+ * hearthParseWireValue() (the same 64-bit-bit-pattern parser already used
+ * for +MTATTR's identical width problem, this file's own include below):
+ * a command payload field can now genuinely carry Matter's full int64
+ * range (PowerAdjustRequest's `power` field is int64 mW), where the
+ * previous uint32_t silently truncated anything past 2^32 -- the exact
+ * defect class this codebase's own AT+MTATTR/AT+MTMEAS pipelines were
+ * widened against earlier (0.8.0). Every existing consumer narrows its own
+ * read of `value[i]` down to a uint8_t/uint16_t/uint32_t/bool with an
+ * explicit or implicit cast, so this widening is behaviour-preserving for
+ * all of them (two's-complement truncation is identical either way); only
+ * a genuinely 64-bit-wide payload field, DEM's `power` being the first,
+ * needs the extra range.
  */
 #define HEARTH_CMD_FIELDS_MAX 5
 struct HearthCmdFields {
   uint8_t count;
   bool present[HEARTH_CMD_FIELDS_MAX];
-  uint32_t value[HEARTH_CMD_FIELDS_MAX];
+  int64_t value[HEARTH_CMD_FIELDS_MAX];
 };
 
 class MatterEndPoint {
